Revert "Complate timer"
This reverts commit 910168d32ad3f97c3eaf58cd8097a8d2e4a0171c.

+++ app/src/main/java/com/o4x/musical/util/PreferenceUtil.kt
@@ -289,11 +289,11 @@ object PreferenceUtil {
         }
 
     @JvmStatic
-    var lastSleepTimerValue: Long
-        get() = sharedPreferences.getLong(LAST_SLEEP_TIMER_VALUE, 30)
+    var lastSleepTimerValue: Int
+        get() = sharedPreferences.getInt(LAST_SLEEP_TIMER_VALUE, 30)
         set(value) {
             val editor = sharedPreferences.edit()
-            editor.putLong(LAST_SLEEP_TIMER_VALUE, value)
+            editor.putInt(LAST_SLEEP_TIMER_VALUE, value)
             editor.apply()
         }
